tools/kustolive/requirements.txt
adal==1.2.2
asn1crypto==0.24.0
azure-common==1.1.23
azure-kusto-data==0.0.35
azure-kusto-ingest==0.0.35
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-queue==2.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
numpy==1.17.2
pandas==0.25.1
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.6
